New storage_api LockFile API

+++ lib/fsblockstore/longtail_fsblockstore.c
@@ -113,62 +113,6 @@ static char* GetTempBlockPath(struct Longtail_StorageAPI* storage_api, const cha
     return storage_api->ConcatPath(storage_api, content_path, file_name);
 }
 
-static Longtail_StorageAPI_HOpenFile LockContentIndex(
-    struct Longtail_StorageAPI* storage_api,
-    HLongtail_SpinLock lock,
-    const char* content_index_lock_path)
-{
-    Longtail_StorageAPI_HOpenFile content_index_lock_file;
-    int try_count = 50;
-    uint64_t retry_delay = 1000;
-    uint64_t total_delay = 0;
-
-    int err = EnsureParentPathExists(storage_api, content_index_lock_path);
-    if (err)
-    {
-        LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "LockContentIndex(%p, `%s`) failed with %d",
-            storage_api, content_index_lock_path,
-            err)
-        return 0;
-    }
-
-    // We try to open the file for write, if we succeed no live process is in the process of updating the index
-    err = storage_api->OpenWriteFile(storage_api, content_index_lock_path, 0, &content_index_lock_file);
-    while (err)
-    {
-        if (lock)
-        {
-            Longtail_UnlockSpinLock(lock);
-        }
-        Longtail_Sleep(retry_delay);
-        total_delay += retry_delay;
-        err = storage_api->OpenWriteFile(storage_api, content_index_lock_path, 0, &content_index_lock_file);
-        if (lock)
-        {
-            Longtail_LockSpinLock(lock);
-        }
-        if (--try_count == 0)
-        {
-            LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "LockContentIndex(%p, `%s`) failed with %d, waited %f seconds",
-                storage_api, content_index_lock_path,
-                err,
-                (total_delay / 1000) / 1000.f)
-            return 0;
-        }
-        retry_delay += 2000;
-    }
-    return content_index_lock_file;
-}
-
-static void UnlockContentIndex(
-    struct Longtail_StorageAPI* storage_api,
-    const char* content_index_lock_path,
-    Longtail_StorageAPI_HOpenFile content_index_lock_file)
-{
-    storage_api->CloseFile(storage_api, content_index_lock_file);
-    storage_api->RemoveFile(storage_api, content_index_lock_path);
-}
-
 static int SafeWriteContentIndex(struct FSBlockStoreAPI* api)
 {
     struct Longtail_StorageAPI* storage_api = api->m_StorageAPI;
@@ -181,24 +125,11 @@ static int SafeWriteContentIndex(struct FSBlockStoreAPI* api)
     int err = EnsureParentPathExists(storage_api, content_index_path_tmp);
     if (err)
     {
-        LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "SafeWriteContentIndex(%p) failed with %d",
+        LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "SafeWriteContentIndex(%p) EnsureParentPathExists() failed with %d",
             api,
             err)
         Longtail_Free((char*)content_index_path_tmp);
         return err;
-    }
-
-    Longtail_StorageAPI_HOpenFile content_index_lock_file = LockContentIndex(
-        storage_api,
-        api->m_Lock,
-        api->m_ContentIndexLockPath);
-    if (!content_index_lock_file)
-    {
-        LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "SafeWriteContentIndex(%p) failed with %d",
-            api,
-            EPERM)
-        Longtail_Free((void*)content_index_path_tmp);
-        return EPERM;
     }
 
     const char* content_index_path = storage_api->ConcatPath(storage_api, content_path, "store.lci");
@@ -223,12 +154,11 @@ static int SafeWriteContentIndex(struct FSBlockStoreAPI* api)
             &merged_content_index);
         if (err)
         {
-            LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "SafeWriteContentIndex(%p) failed with %d",
+            LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "SafeWriteContentIndex(%p) Longtail_MergeContentIndex() failed with %d",
                 api,
                 err)
             Longtail_Free(existing_content_index);
             Longtail_Free((void*)content_index_path);
-            UnlockContentIndex(storage_api, api->m_ContentIndexLockPath, content_index_lock_file);
             Longtail_Free((void*)content_index_path_tmp);
             return err;
         }
@@ -238,12 +168,11 @@ static int SafeWriteContentIndex(struct FSBlockStoreAPI* api)
     err = Longtail_WriteContentIndex(storage_api, content_index, content_index_path_tmp);
     if (err)
     {
-        LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "SafeWriteContentIndex(%p) failed with %d",
+        LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "SafeWriteContentIndex(%p) Longtail_WriteContentIndex() failed with %d",
             api,
             err)
         Longtail_Free(existing_content_index);
         Longtail_Free((void*)content_index_path);
-        UnlockContentIndex(storage_api, api->m_ContentIndexLockPath, content_index_lock_file);
         Longtail_Free((void*)content_index_path_tmp);
         return err;
     }
@@ -251,15 +180,14 @@ static int SafeWriteContentIndex(struct FSBlockStoreAPI* api)
     if (storage_api->IsFile(storage_api, content_index_path))
     {
         err = storage_api->RemoveFile(storage_api, content_index_path);
-        if (err)
+        if (err && err != ENOENT)
         {
-            LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "SafeWriteContentIndex(%p) failed with %d",
+            LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "SafeWriteContentIndex(%p) RemoveFile() failed with %d",
                 api,
                 err)
             Longtail_Free(existing_content_index);
             Longtail_Free((void*)content_index_path);
             storage_api->RemoveFile(storage_api, content_index_path_tmp);
-            UnlockContentIndex(storage_api, api->m_ContentIndexLockPath, content_index_lock_file);
             Longtail_Free((void*)content_index_path_tmp);
             return err;
         }
@@ -268,19 +196,20 @@ static int SafeWriteContentIndex(struct FSBlockStoreAPI* api)
     err = storage_api->RenameFile(storage_api, content_index_path_tmp, content_index_path);
     if (err)
     {
-        LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "SafeWriteContentIndex(%p) failed with %d",
+        LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "SafeWriteContentIndex(%p) RenameFile() failed with %d",
             api,
             err)
         storage_api->RemoveFile(storage_api, content_index_path_tmp);
     }
 
-    if (api->m_ContentIndex != content_index)
+    if (!err)
     {
-        Longtail_Free(api->m_ContentIndex);
-        api->m_ContentIndex = content_index;
+        if (api->m_ContentIndex != content_index)
+        {
+            Longtail_Free(api->m_ContentIndex);
+            api->m_ContentIndex = content_index;
+        }
     }
-
-    UnlockContentIndex(storage_api, api->m_ContentIndexLockPath, content_index_lock_file);
 
     Longtail_Free(existing_content_index);
     Longtail_Free((void*)content_index_path);
@@ -802,16 +731,22 @@ int FSBlockStore_GetContentIndexFromStorage(
 
     struct Longtail_ContentIndex* content_index = 0;
 
-    Longtail_StorageAPI_HOpenFile content_index_lock_file = LockContentIndex(
-        fsblockstore_api->m_StorageAPI,
-        fsblockstore_api->m_Lock,
-        fsblockstore_api->m_ContentIndexLockPath);
-    if (!content_index_lock_file)
+    int err = EnsureParentPathExists(storage_api, fsblockstore_api->m_ContentIndexLockPath);
+    if (err)
     {
         LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "FSBlockStore_GetContentIndexFromStorage(%p, %p) failed with %d",
             fsblockstore_api, out_content_index,
-            EPERM)
-        return EPERM;
+            err)
+        return err;
+    }
+    Longtail_StorageAPI_HLockFile content_index_lock_file;
+    err = storage_api->LockFile(storage_api, fsblockstore_api->m_ContentIndexLockPath, &content_index_lock_file);
+    if (err)
+    {
+        LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "FSBlockStore_GetContentIndexFromStorage(%p, %p) failed with %d",
+            fsblockstore_api, out_content_index,
+            err)
+        return err;
     }
 
     const char* content_index_path = storage_api->ConcatPath(storage_api, content_path, "store.lci");
@@ -825,11 +760,11 @@ int FSBlockStore_GetContentIndexFromStorage(
                 fsblockstore_api, out_content_index,
                 err)
             Longtail_Free((void*)content_index_path);
-            UnlockContentIndex(fsblockstore_api->m_StorageAPI, fsblockstore_api->m_ContentIndexLockPath, content_index_lock_file);
+            storage_api->UnlockFile(storage_api, content_index_lock_file);
             return err;
         }
     }
-    UnlockContentIndex(fsblockstore_api->m_StorageAPI, fsblockstore_api->m_ContentIndexLockPath, content_index_lock_file);
+    storage_api->UnlockFile(storage_api, content_index_lock_file);
 
     Longtail_Free((void*)content_index_path);
     if (content_index)
@@ -837,7 +772,7 @@ int FSBlockStore_GetContentIndexFromStorage(
         *out_content_index = content_index;
         return 0;
     }
-    int err = ReadContent(
+    err = ReadContent(
         storage_api,
         job_api,
         default_max_block_size,
@@ -1073,16 +1008,32 @@ static int FSBlockStore_Flush(struct Longtail_BlockStoreAPI* block_store_api, st
 
     if (api->m_ContentIndex)
     {
-        const char* content_index_path = api->m_StorageAPI->ConcatPath(api->m_StorageAPI, api->m_ContentPath, "store.lci");
-        if ((!api->m_StorageAPI->IsFile(api->m_StorageAPI, content_index_path)) || (new_block_count > 0))
+        int err = EnsureParentPathExists(api->m_StorageAPI, api->m_ContentIndexLockPath);
+        if (err)
         {
-            err = SafeWriteContentIndex(api);
-            if (err)
+            LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_ERROR, "FSBlockStore_Flush(%p, %p) failed with %d",
+                block_store_api, async_complete_api,
+                err)
+        }
+        else
+        {
+            Longtail_StorageAPI_HLockFile content_index_lock_file;
+            int err = api->m_StorageAPI->LockFile(api->m_StorageAPI, api->m_ContentIndexLockPath, &content_index_lock_file);
+            if (!err)
             {
-                LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_WARNING, "Failed to store content index for `%s`, %d", api->m_ContentPath, err);
+                const char* content_index_path = api->m_StorageAPI->ConcatPath(api->m_StorageAPI, api->m_ContentPath, "store.lci");
+                if ((!api->m_StorageAPI->IsFile(api->m_StorageAPI, content_index_path)) || (new_block_count > 0))
+                {
+                    err = SafeWriteContentIndex(api);
+                    if (err)
+                    {
+                        LONGTAIL_LOG(LONGTAIL_LOG_LEVEL_WARNING, "Failed to store content index for `%s`, %d", api->m_ContentPath, err);
+                    }
+                }
+                Longtail_Free((void*)content_index_path);
+                api->m_StorageAPI->UnlockFile(api->m_StorageAPI, content_index_lock_file);
             }
         }
-        Longtail_Free((void*)content_index_path);
     }
 
     Longtail_UnlockSpinLock(api->m_Lock);
@@ -1166,7 +1117,7 @@ static int FSBlockStore_Init(
     api->m_BlockState = 0;
     api->m_AddedBlockIndexes = 0;
     api->m_BlockExtension = optional_extension ? optional_extension : ".lrb";
-    api->m_ContentIndexLockPath = storage_api->ConcatPath(storage_api, content_path, "store.lci.lock");
+    api->m_ContentIndexLockPath = storage_api->ConcatPath(storage_api, content_path, "store.lci.sync");
 
     GetUniqueExtension(unique_id, api->m_TmpExtension);
     api->m_DefaultMaxBlockSize = default_max_block_size;
